// pdlearning/src/web-apps/src/libraries/domain-api/src/course/models/learner-violation-course-criteria-type.model.ts
export enum CourseCriteriaLearnerViolationType {
  NotViolate = 'NotViolate',
  Missing = 'Missing',
  OverMaxParticipant = 'OverMaxParticipant'
}

export enum CourseCriteriaLearnerViolationField {
  AccountType = 'AccountType',
  ServiceSchemes = 'ServiceSchemes',
  Tracks = 'Tracks',
  DevRoles = 'DevRoles',
  TeachingLevels = 'TeachingLevels',
  TeachingCourseOfStudy = 'TeachingCourseOfStudy',
  JobFamily = 'JobFamily',
  CoCurricularActivity = 'CoCurricularActivity',
  SubGradeBanding = 'SubGradeBanding',
  PlaceOfWork = 'PlaceOfWork',
  PreRequisiteCourses = 'PreRequisiteCourses'
}